python/sqlmodel-fastapi/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["manupanand <manupanandindianinvestor@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.6"
sqlmodel = "^0.0.22"
python-dotenv = "^1.0.1"
psycopg = "^3.2.3"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
